requirements.txt
pandas
numpy
torch
transformers
faiss-cpu
gradio
tqdm
openpyxl
pathlib
logging